Refactor attendance summary handling and improve precision
- Changed HTTP method for generating monthly summaries to POST in AttendanceController.
- Removed MonthlyAttendanceSummaryController for restructuring.
- Updated PayrollController to use decimal for tax and net salary calculations.
- Added optional EmployeeFullName property in EmployeeAttendanceDto.
- Set precision for payroll-related properties in GuesthouseDbContext.
- Changed relationship between MonthlyAttendanceSummary and Employee to one-to-many.
- Updated MonthlyAttendanceSummaryInterface method signatures.
- Reflected database structure changes in GuesthouseDbContextModelSnapshot.
- Enhanced MonthlyAttendanceSummaryRepo for new summary logic and transaction handling.
- Updated UI Attendance.cs to align with API changes.
- Modified UI forms to include controls for generating and displaying monthly summaries.
- Created migration files for database schema updates.

+++ SGH_API/Controllers/PayrollController.cs
@@ -89,9 +89,9 @@ namespace SolmileGuesthouseAPI.Controllers
                     return BadRequest("No applicable tax bracket found for the given basic salary.");
                 }
 
-                payroll.Tax = (decimal)(payroll.BasicSalary * 0.15); 
+                payroll.Tax = payroll.BasicSalary * 0.15m; 
 
-                payroll.NetSalary = (double)payroll.BasicSalary + (double)payroll.Allowances - (double)payroll.Deductions - (double)payroll.Tax;
+                payroll.NetSalary = payroll.BasicSalary + payroll.Allowances - payroll.Deductions - payroll.Tax;
 
 
                 var createdPayroll = await _payrollInterface.CreateOrUpdatePayrollAsync(payroll);
