+++ SharedModel/Models/Payroll.cs
@@ -4,11 +4,12 @@
     {
         public int PayrollId { get; set; }
         public int EmployeeId { get; set; }
-        public double BasicSalary { get; set; }
-        public double Allowances { get; set; }
+
+        public decimal BasicSalary { get; set; }
+        public decimal Allowances { get; set; }
         public decimal Tax { get; set; }
-        public double Deductions { get; set; }
-        public double NetSalary { get; set; }
+        public decimal Deductions { get; set; }
+        public decimal NetSalary { get; set; }
         public DateTime PayPeriod { get; set; }
         public virtual Employee Employee { get; set; }
         public string DeductionReason { get; set; }
